Remove redundant expression from file names
