+++ src/contact_constitutivelaw/src/4C_contact_constitutivelaw_brokenrational.hpp
@@ -5,8 +5,8 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#ifndef FOUR_C_CONTACT_CONSTITUTIVELAW_BROKENRATIONAL_CONTACTCONSTITUTIVELAW_HPP
-#define FOUR_C_CONTACT_CONSTITUTIVELAW_BROKENRATIONAL_CONTACTCONSTITUTIVELAW_HPP
+#ifndef FOUR_C_CONTACT_CONSTITUTIVELAW_BROKENRATIONAL_HPP
+#define FOUR_C_CONTACT_CONSTITUTIVELAW_BROKENRATIONAL_HPP
 
 
 #include "4C_config.hpp"
