+++ src/contact_constitutivelaw/src/4C_contact_constitutivelaw_contactconstitutivelaw.cpp
@@ -7,18 +7,18 @@
 
 #include "4C_contact_constitutivelaw_contactconstitutivelaw.hpp"
 
-#include "4C_contact_constitutivelaw_brokenrational_contactconstitutivelaw.hpp"
+#include "4C_contact_constitutivelaw_brokenrational.hpp"
 #include "4C_contact_constitutivelaw_bundle.hpp"
 #include "4C_contact_constitutivelaw_contactconstitutivelaw_parameter.hpp"
-#include "4C_contact_constitutivelaw_cubic_contactconstitutivelaw.hpp"
-#include "4C_contact_constitutivelaw_linear_contactconstitutivelaw.hpp"
-#include "4C_contact_constitutivelaw_power_contactconstitutivelaw.hpp"
-#include "4C_contact_constitutivelaw_python_surrogate_contactconstitutivelaw.hpp"
+#include "4C_contact_constitutivelaw_cubic.hpp"
+#include "4C_contact_constitutivelaw_linear.hpp"
+#include "4C_contact_constitutivelaw_power.hpp"
+#include "4C_contact_constitutivelaw_python_surrogate.hpp"
 #include "4C_global_data.hpp"
 #include "4C_utils_enum.hpp"
 
 #ifdef FOUR_C_WITH_MIRCO
-#include "4C_contact_constitutivelaw_mirco_contactconstitutivelaw.hpp"
+#include "4C_contact_constitutivelaw_mirco.hpp"
 #endif
 
 FOUR_C_NAMESPACE_OPEN
