+++ src/contact_constitutivelaw/src/4C_contact_constitutivelaw_mirco.hpp
@@ -5,8 +5,8 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#ifndef FOUR_C_CONTACT_CONSTITUTIVELAW_MIRCO_CONTACTCONSTITUTIVELAW_HPP
-#define FOUR_C_CONTACT_CONSTITUTIVELAW_MIRCO_CONTACTCONSTITUTIVELAW_HPP
+#ifndef FOUR_C_CONTACT_CONSTITUTIVELAW_MIRCO_HPP
+#define FOUR_C_CONTACT_CONSTITUTIVELAW_MIRCO_HPP
 
 #include "4C_config.hpp"
