+++ src/contact_constitutivelaw/src/4C_contact_constitutivelaw_power.cpp
@@ -5,18 +5,20 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_contact_constitutivelaw_linear_contactconstitutivelaw.hpp"
+#include "4C_contact_constitutivelaw_power.hpp"
 
 #include "4C_global_data.hpp"
 #include "4C_linalg_serialdensematrix.hpp"
 #include "4C_linalg_serialdensevector.hpp"
+
+#include <math.h>
 
 FOUR_C_NAMESPACE_OPEN
 
 
 /*----------------------------------------------------------------------*/
 /*----------------------------------------------------------------------*/
-CONTACT::CONSTITUTIVELAW::LinearConstitutiveLawParams::LinearConstitutiveLawParams(
+CONTACT::CONSTITUTIVELAW::PowerConstitutiveLawParams::PowerConstitutiveLawParams(
     const Core::IO::InputParameterContainer& container)
     : CONTACT::CONSTITUTIVELAW::Parameter(container),
       a_(container.get<double>("A")),
@@ -26,15 +28,15 @@ CONTACT::CONSTITUTIVELAW::LinearConstitutiveLawParams::LinearConstitutiveLawPara
 
 /*----------------------------------------------------------------------*/
 /*----------------------------------------------------------------------*/
-CONTACT::CONSTITUTIVELAW::LinearConstitutiveLaw::LinearConstitutiveLaw(
-    CONTACT::CONSTITUTIVELAW::LinearConstitutiveLawParams params)
+CONTACT::CONSTITUTIVELAW::PowerConstitutiveLaw::PowerConstitutiveLaw(
+    CONTACT::CONSTITUTIVELAW::PowerConstitutiveLawParams params)
     : params_(std::move(params))
 {
 }
 
 /*----------------------------------------------------------------------*
  *----------------------------------------------------------------------*/
-double CONTACT::CONSTITUTIVELAW::LinearConstitutiveLaw::evaluate(
+double CONTACT::CONSTITUTIVELAW::PowerConstitutiveLaw::evaluate(
     const double gap, CONTACT::Node* cnode)
 {
   if (gap + params_.get_offset() > 0.0)
@@ -45,12 +47,19 @@ double CONTACT::CONSTITUTIVELAW::LinearConstitutiveLaw::evaluate(
         gap, params_.get_offset());
   }
 
-  return params_.getdata() * (gap + params_.get_offset()) + params_.get_b();
+  const double result = -(params_.getdata() * pow(-gap - params_.get_offset(), params_.get_b()));
+
+  if (result > 0)
+    FOUR_C_THROW(
+        "The constitutive function you are using seems to be positive, even though the gap is "
+        "negative. Please check your coefficients!");
+
+  return result;
 }
 
 /*----------------------------------------------------------------------*
  *----------------------------------------------------------------------*/
-double CONTACT::CONSTITUTIVELAW::LinearConstitutiveLaw::evaluate_derivative(
+double CONTACT::CONSTITUTIVELAW::PowerConstitutiveLaw::evaluate_derivative(
     const double gap, CONTACT::Node* cnode)
 {
   if (gap + params_.get_offset() > 0.0)
@@ -61,7 +70,8 @@ double CONTACT::CONSTITUTIVELAW::LinearConstitutiveLaw::evaluate_derivative(
         gap, params_.get_offset());
   }
 
-  return params_.getdata();
+  return params_.getdata() * params_.get_b() *
+         pow(-gap - params_.get_offset(), params_.get_b() - 1);
 }
 
 FOUR_C_NAMESPACE_CLOSE
